% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{AR1_MODEL$..init..}
\alias{AR1_MODEL$..init..}
\title{Initialization method}
\arguments{
\item{phi}{Numeric. The parameter for the \code{AR(1)} term. Default is \code{phi = 0.5}.}

\item{sigma}{Positive numeric. Default is \code{sigma = 1}.}

\item{x}{Random variable or closed form expression. Default is
\code{x = rand_uniform(-1, 1, env = new.env(parent = parent.env(self)))}.}

\item{e}{Random variable or closed form expression. Default is
\code{e = rand_normal(0, sigma, env = new.env(parent = parent.env(self)))}.}
}
\value{
Return the object itself.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment. The response variable of this model
is \code{y}. The formula of y is defined in \link{AR1_MODEL$formula}, the null
formula is defined in \link{AR1_MODEL$null_formula}, the alternative is
defined in \link{AR1_MODEL$alt_formula}.
}
\examples{

# Instantiate
x <- rand_uniform()
e <- rand_normal(sigma = 0.5)

test <- ar1_model(phi = 0.9, x = x, e = e)

test

# Generate data
test$gen(10)

# Generate lineup
test$gen_lineup(10, k = 3)

# Plot the lineup
test$plot_lineup(test$gen_lineup(100))

test <- ar1_model(phi = 0.1, x = x, e = e)
test$plot_lineup(test$gen_lineup(100))

test <- ar1_model(phi = 0.3, x = x, e = e)
test$plot_lineup(test$gen_lineup(100))

test <- ar1_model(phi = 0.5, x = x, e = e)
test$plot_lineup(test$gen_lineup(100))

test <- ar1_model(phi = 0.7, x = x, e = e)
test$plot_lineup(test$gen_lineup(100))
}
