% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\docType{data}
\name{AR1_MODEL}
\alias{AR1_MODEL}
\title{AR1_MODEL class environment}
\format{
An environment with S3 class \code{bandicoot_oop}.
}
\usage{
AR1_MODEL
}
\description{
This is the class of visual inference orthogonal polynomial
linear model, inherited from \link{VI_MODEL}.
}
\seealso{
Parent class: \link{VI_MODEL}
\cr
\cr
New attributes: \link{AR1_MODEL$formula},
\link{AR1_MODEL$null_formula}, \link{AR1_MODEL$alt_formula}
\cr
\cr
New methods: \link{AR1_MODEL$..init..}, \link{AR1_MODEL$test}, \link{AR1_MODEL$set_prm}, \link{AR1_MODEL$ar1}
}
\keyword{datasets}
