% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{BASE$..new..}
\alias{BASE$..new..}
\title{Build a new instance from a class or an instance}
\arguments{
\item{env}{Environment. The instance environment.}

\item{init_call}{Call. Contents of the \code{..init_call..}. It is recommended to
use the default value.}
}
\value{
An instance environment.
}
\description{
This function will copy all methods and attributes, except
the container, and the instantiation method. Then, the \code{..init_call..} attribute
will be set to the current system call, and the \code{..instantiated..} attribute
will be set to \code{TRUE}. Notice, the \code{..init..} method will not run.
}
\examples{

BASE$..new..()

# Inherit from BASE
TEST <- new_class(BASE, class_name = "TEST")

TEST$..new..()
}
