% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{BASE$instantiation}
\alias{BASE$instantiation}
\title{Instantiation method}
\arguments{
\item{...}{Arguments passed to \code{..init..} method.}

\item{env}{Environment. The instance environment.}
}
\value{
An instance environment.
}
\description{
This function will new an instance using the \code{..new..} method,
then initialized the instance with the \code{..init..} method.
}
\examples{

BASE$..dict..()

# Build an instance
base_instance <- BASE$instantiation()

base_instance$..dict..()
}
