% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{BASE$set_attr}
\alias{BASE$set_attr}
\title{Set value of an attribute or a method}
\arguments{
\item{attr_name}{Character. Attribute name.}

\item{attr_val}{Any value.}
}
\value{
No return value, called for side effects.
}
\description{
This function sets the value of an attribute or a method.
}
\examples{

BASE$set_attr("test", 1)

BASE$test
}
