% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_BASE.R
\docType{data}
\name{BASE}
\alias{BASE}
\title{BASE class environment}
\format{
An environment with S3 class \code{oop}.
}
\usage{
BASE
}
\description{
This class provides essential attributes and methods.
}
\keyword{datasets}
