% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{calc_p_value}
\alias{calc_p_value}
\title{Calculate p-value for a visual test}
\usage{
calc_p_value(
  n_detect,
  n_eval,
  n_sel,
  n_plot = 20,
  n_sim = 50000,
  cache_env = NULL,
  dist = "uniform",
  alpha = 1
)
}
\arguments{
\item{n_detect}{Integer. Observed value of number of detections.}

\item{n_eval}{Integer. Number of evaluations.}

\item{n_sel}{Integer. A vector of the number of selections.}

\item{n_plot}{Integer. Number of plots in the lineup.}

\item{n_sim}{Integer. Number of simulations draws.}

\item{cache_env}{Environment. A provided environment for caching.}

\item{dist}{Character. Name of the distribution used for the attractiveness simulation.
One of "uniform" and "dirichlet".}

\item{alpha}{Numeric. A single parameter value used by the Dirichlet distribution.}
}
\value{
A numeric value representing the p-value.
}
\description{
This function calculate the p-value for a visual test. The null distribution
is simulated by using function \code{\link[=sim_dist]{sim_dist()}}. The p-value is the probability
of the number of detections greater or equal to the observed value.
}
\details{
It is encouraged to provide a cache environment to boost up the performance
when this function needs to be reused. The cache environment will remember
the result corresponding to the combinations of \code{n_eval}
and \code{n_sim}.

There are two attractiveness distribution available,
one is uniform distribution, another is Dirichlet distribution. Uniform
distribution ensures the marginal distribution of the probability of every
plot being selected is uniform.
When \eqn{\alpha = 1},
Dirichlet distribution ensures the probability of every plot being selected
is evenly distributed in a standard \eqn{M - 1} simplex.
}
\examples{
calc_p_value(2, 3, c(1,1,2))
calc_p_value(1, 1, c(1))

}
