% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{calc_p_value_comb}
\alias{calc_p_value_comb}
\title{Calculate p-value for all combinations of evaluations of a lineup.}
\usage{
calc_p_value_comb(
  detected,
  n_eval,
  n_sel,
  n_plot = 20,
  n_sim = 50000,
  cache_env = NULL
)
}
\arguments{
\item{detected}{Boolean. A vector of Boolean values indicating whether the
lineup has been detected by the subjects.}

\item{n_eval}{Integer. Desired number of evaluations.}

\item{n_sel}{Integer. A vector of the number of selections.}

\item{n_plot}{Integer. Number of plots in the lineup.}

\item{n_sim}{Integer. Number of simulation draws.}

\item{cache_env}{Environment. A provided environment for caching.}
}
\value{
A vector of p-value with the combination matrix as the attribute.
}
\description{
This function calculate p-value for all combinations of evaluations given
the desired number of evaluations. For a lineup that being evaluated
by \eqn{K} subjects, all combinations of \eqn{K} evaluations, taken
\eqn{X} at a time will be generated, where \eqn{X} is the desired number of
evaluations.
}
\details{
It is encouraged to provide a cache environment to boost up the performance
when this function needs to be reused. The cache environment will remember
the result corresponding to the combinations of \code{n_eval}
and \code{n_sim}.
}
\examples{
calc_p_value_comb(c(TRUE, FALSE, TRUE), 2, c(1, 1, 2))
calc_p_value_comb(c(TRUE, FALSE), 1, c(1, 1))

}
\seealso{
\code{\link[=combn]{combn()}}
}
