% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{class_method}
\alias{class_method}
\title{Get the class method}
\usage{
class_method(env, cls, method_name, ..., container_name = "method_env_")
}
\arguments{
\item{env}{Environment. The instance environment.}

\item{cls}{Function. The class constructor.}

\item{method_name}{Character. The method name.}

\item{...}{Arguments passed to the class constructor.}

\item{container_name}{Character. Container name of the instance environment.}
}
\value{
The class method.
}
\description{
This function construct an instance by using the provided class constructor,
then get the method from the instance and set its evaluation environment to
the container of the provided instance environment.
}
\details{
If the body of the method depends on the arguments passed to the class
constructor, those values needs to be provided in \code{...}.
}
