% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_CLOSED_FORM.R
\name{CLOSED_FORM$..len..}
\alias{CLOSED_FORM$..len..}
\title{Length of the object}
\value{
An integer.
}
\description{
This function recursively count symbols
stored in the closed form expression and nested closed form expression.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{CLOSED_FORM$..len..()
}\if{html}{\out{</div>}}
}
}
\examples{

test <- closed_form(~1+1)
test$..len..()
}
