% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_CLOSED_FORM.R
\name{CLOSED_FORM$gen_rhs}
\alias{CLOSED_FORM$gen_rhs}
\title{Generating random values from the expression with right hand side being kept}
\arguments{
\item{n}{Integer. Number of observations.}
}
\value{
Numeric values.
}
\description{
This function generates random values from the expression
with right hand side being kept. Random values will be generated
independently as long as they don't share the same symbol in the expression.
}
\seealso{
\link{CLOSED_FORM$gen}
}
