% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_CLOSED_FORM.R
\name{CLOSED_FORM$update_expr}
\alias{CLOSED_FORM$update_expr}
\title{Update the closed form expression}
\arguments{
\item{expr}{Formula. Only the right hand side of the last \code{~} will be kept as
the final expression.}
}
\value{
No return value, called for side effects.
}
\description{
This function updates the closed form expression. It will not
update the symbol values. For updating symbol values,
please check \link{CLOSED_FORM$update_sym}.
}
\examples{

# Constant variable
a <- 1

# Random uniform variable
z <- closed_form(~2 + a)

z$compute()

# Update the expression
z$update_expr(~2 - a)

z$compute()
}
