% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_CLOSED_FORM.R
\name{CLOSED_FORM}
\alias{CLOSED_FORM}
\alias{closed_form}
\title{CLOSED_FORM class environment}
\usage{
closed_form(
  expr,
  env = new.env(parent = parent.frame()),
  init_call = sys.call()
)
}
\arguments{
\item{expr}{Formula. Only the right hand side of the last \code{~} will be kept as
the final expression.}

\item{env}{Environment. The instance environment.}

\item{init_call}{Call. Contents of the \code{..init_call...} It is recommended to leave it
as default.}
}
\value{
An instance environment.
}
\description{
This is the base class of closed form expression, inherited from
\link[bandicoot:BASE]{bandicoot::BASE}. It is an environment with S3 class \code{bandicoot_oop}.\cr\cr
For instantiation, any simple expressions can be provided,
as long as all the symbols exist in the current environment. Note that it
tries to evaluate \strong{ALL the atomic symbols} in the expression during
initialization, and store the values in the object. Hence, calls like
\code{a$b} will also be decomposed as \code{$}, \code{a} and \code{b}, where \code{b} will be
interpreted as \strong{a variable "b" exists in the current environment}. Therefore,
use \code{~a[["b"]]} instead of \code{~a$b}. And pre-define function like
\code{myfun = function() 1}, then use it in the expression \code{~myfun()}.
}
\section{Functions}{
\itemize{
\item \code{closed_form()}: Class constructor, same as \code{CLOSED_FORM$instantiate()}.

}}
\section{Class information}{
\subsection{Parent classes}{
\itemize{
\item Direct:
\itemize{
\item \link[bandicoot:BASE]{bandicoot::BASE}
}
}
}

\subsection{New attributes}{
\itemize{
\item E:
\itemize{
\item \link{CLOSED_FORM$expr}
}
\item S:
\itemize{
\item \link{CLOSED_FORM$sym}
\item \link{CLOSED_FORM$sym_name}
\item \link{CLOSED_FORM$sym_type}
}
}
}

\subsection{New methods}{
\itemize{
\item A:
\itemize{
\item \link{CLOSED_FORM$as_dataframe}
\item \link{CLOSED_FORM$ast}
}
\item C:
\itemize{
\item \link{CLOSED_FORM$compute}
}
\item G:
\itemize{
\item \link{CLOSED_FORM$gen}
}
\item I:
\itemize{
\item \link{CLOSED_FORM$..init..}
}
\item L:
\itemize{
\item \link{CLOSED_FORM$..len..}
}
\item S:
\itemize{
\item \link{CLOSED_FORM$..str..}
}
}
}
}

