% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{CUBIC_MODEL$effect_size}
\alias{CUBIC_MODEL$effect_size}
\title{Compute the effect size of the simulated data}
\arguments{
\item{dat}{Dataframe/List. List contains variable \code{x} and \code{z}.}

\item{a}{Numeric. Default is \code{a = self$prm$a}.}

\item{b}{Numeric. Default is \code{b = self$prm$b}.}

\item{c}{Numeric. Default is \code{c = self$prm$c}.}

\item{sigma}{Positive numeric. Default is \code{sigma = self$prm$sigma}.}
}
\value{
A numeric value.
}
\description{
This function computes the effect size of the simulated data.
}
\examples{

mod <- cubic_model(-1, 1, 1, 0.5)
dat <- mod$gen(1000, fit_model = TRUE)
mod$effect_size(dat)
}
