% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{CUBIC_MODEL$formula}
\alias{CUBIC_MODEL$formula}
\title{Closed form expression of \code{y}}
\description{
A quoted formula, will be passed to \code{CLOSED_FORM$instantiate} to
define a closed form expression for \code{y}.
}
\examples{

CUBIC_MODEL$formula
}
