% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\docType{data}
\name{CUBIC_MODEL}
\alias{CUBIC_MODEL}
\title{CUBIC_MODEL class environment}
\format{
An environment with S3 class \code{visage_oop}.
}
\usage{
CUBIC_MODEL
}
\description{
This is the class of visual inference cubic linear model,
inherited from \link{VI_MODEL}.
}
\seealso{
Parent class: \link{VI_MODEL}
\cr
\cr
New attributes: \link{CUBIC_MODEL$formula},
\link{CUBIC_MODEL$null_formula}, \link{CUBIC_MODEL$alt_formula},
\cr
\cr
New methods: \link{CUBIC_MODEL$..init..}, \link{CUBIC_MODEL$E},
\link{CUBIC_MODEL$effect_size}
}
\keyword{datasets}
