% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{define_pkg_fns}
\alias{define_pkg_fns}
\title{Load functions from package namespaces into current environment}
\usage{
define_pkg_fns(..., pkg = NULL)
}
\arguments{
\item{...}{Functions. Preferred names can be provide via named arguments.}

\item{pkg}{Package.}
}
\value{
No return value, called for side effects
}
\description{
This function loads functions from package namespaces via \code{::} and
assigns them to the preferred function names in the current environment.
}
\details{
Preferred function names can be provide via named arguments
like \code{dplyr_filter = filter}.
}
\examples{
define_pkg_fns(select, filter, pkg = dplyr)
define_pkg_fns(select, dplyr_filter = filter, `\%>\%`, pkg = dplyr)

}
