% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{define_pkg_funcs}
\alias{define_pkg_funcs}
\title{Load functions from package namespaces into current environment}
\usage{
define_pkg_funcs(func_names, pkg_names, pref_names = func_names)
}
\arguments{
\item{func_names}{Character. A vector of function names.}

\item{pkg_names}{Character. A vector of package names.}

\item{pref_names}{Character. A vector of preferred function names.}
}
\value{
No return value, called for side effects
}
\description{
This function loads functions from package namespaces via \code{::} and
assigns them to the preferred function names in the current environment.
}
\details{
If the length of \code{pref_names} and \code{func_names} does not match, \code{func_names}
will repleace \code{pref_names}.
}
\examples{
define_pkg_funcs("filter", "dplyr")
define_pkg_funcs(c("filter", "\%>\%"), rep("dplyr", 2))

}
