% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{evaluate_p_value}
\alias{evaluate_p_value}
\title{Evaluate test for given p-value and significance level}
\usage{
evaluate_p_value(p_value, significance_level = 0.05, tol = 1e-06)
}
\arguments{
\item{p_value}{Numeric. A vector of p-values.}

\item{significance_level}{Numeric. The significance level.}

\item{tol}{Numeric. Tolerance.}
}
\value{
A vector of Boolean values indicating whether the null hypotheses
should be rejected.
}
\description{
This function reject the null hypothesis if the p-value is smaller than the
given significance level plus the tolerance.
}
\examples{

evaluate_p_value(c(0.05, 0.0501), 0.05, 1e-3)
evaluate_p_value(c(0.05, 0.0501), 0.05, 1e-5)

}
