% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{exact_dist}
\alias{exact_dist}
\title{Calculate the exact distribution of number of detections of a lineup}
\usage{
exact_dist(n_eval, n_plot = 20, dist = "dirichlet", alpha = 1)
}
\arguments{
\item{n_eval}{Integer. Number of evaluations.}

\item{n_plot}{Integer. Number of plots in the lineup.}

\item{dist}{Character. Name of the distribution used for the attractiveness model.
Currently, this can only be "dirichlet".}

\item{alpha}{Numeric. A single parameter value used by the Dirichlet distribution.}
}
\value{
A named vector representing the probability mass function of the distribution.
}
\description{
This function calculates the exact distribution of number of detections of a
lineup assuming every subject selects exactly one plot.
}
\details{
For a given lineup, plots are assumed to have weights
\eqn{W_i, i = 1, ..., M,} where \eqn{M} is the number of plots, and
\eqn{W_i} follows a attractiveness distribution independently. \cr\cr
Currently, this function only supports Dirichlet distribution as
the attractiveness distribution.
Dirichlet distribution ensures the probability of every plot being selected
is evenly distributed in a standard \eqn{M - 1} simplex.
}
\examples{

# The first person select 2 plots, the second one select 2 plots and
# the third one select 3 plots.
sim_dist(c(2,2,3))

}
