% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{HETER_MODEL$alt_formula}
\alias{HETER_MODEL$alt_formula}
\title{Formula for fitting the alternative model}
\description{
Quoted formula for fitting the alternative model.
}
\examples{

HETER_MODEL$alt_formula
}
