% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{HETER_MODEL$sigma_quote_formula}
\alias{HETER_MODEL$sigma_quote_formula}
\title{Quoted formula for constructing \code{sigma}}
\description{
Quoted formula for constructing \code{sigma}.
}
\examples{

HETER_MODEL$sigma_quote_formula
}
