% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{HETER_MODEL$test}
\alias{HETER_MODEL$test}
\title{Test the null model}
\arguments{
\item{dat}{Data frame. A data frame containing all variables needed by the
\code{null_formula}.}

\item{null_formula}{Formula. Formula for fitting the null model. Default
is \code{null_formula = self$null_formula}.}
}
\value{
A list containing the test name, the test statistic and the p-value.
}
\description{
This function test the null model.
In this class, the test is a BP-test computed using \link[lmtest:bptest]{lmtest::bptest}. The
variance formula is \code{~ x + I(x^2)}.
}
\examples{

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- heter_model(a = 0, b = 100)


dat <- test$gen(100000)
test$test(dat)
}
