% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{HIGHER_ORDER_MODEL$..init..}
\alias{HIGHER_ORDER_MODEL$..init..}
\title{Initialization method}
\arguments{
\item{a}{Numeric. Default is \code{a = 1}.}

\item{b}{Numeric. Default is \code{b = 1}.}

\item{c}{Numeric. Default is \code{c = 1}.}

\item{x}{Random variable or closed form expression. Default is
\code{x = rand_uniform(env = new.env(parent = parent.env(self)))}.}

\item{z}{Random variable or closed form expression. Default is
\code{z = rand_uniform(env = new.env(parent = parent.env(self)))}.}

\item{e}{Random variable or closed form expression. Default is
\code{e = rand_normal(env = new.env(parent = parent.env(self)))}.}
}
\value{
No return value, called for side effects.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment. The response variable of this model
is \code{y}. The formula of y is defined in \link{HIGHER_ORDER_MODEL$formula}, the null
formula is defined in \link{HIGHER_ORDER_MODEL$null_formula}, the alternative is
defined in \link{HIGHER_ORDER_MODEL$alt_formula}.
}
\examples{

# Instantiation
x <- rand_uniform()
z <- rand_uniform()
e <- rand_normal()

test <- higher_order_model(a = 200, b = 200, c = 1, x = x, z = z, e = e)

test

# Generate data
test$gen(10)

# Generate lineup
test$gen_lineup(10, k = 3)

# Plot the lineup
test$plot_lineup(test$gen_lineup(100))
}
