% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{HIGHER_ORDER_MODEL$formula}
\alias{HIGHER_ORDER_MODEL$formula}
\title{Closed form expression of \code{y}}
\description{
A formula, will be passed to \code{CLOSED_FORM$instantiation} to
define a closed form expression for \code{y}.
}
\examples{

HIGHER_ORDER_MODEL$formula
}
