% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{HIGHER_ORDER_MODEL$null_formula}
\alias{HIGHER_ORDER_MODEL$null_formula}
\title{Formula for fitting the null model}
\description{
Formula for fitting the null model.
}
\examples{

HIGHER_ORDER_MODEL$null_formula
}
