% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\docType{data}
\name{HIGHER_ORDER_MODEL}
\alias{HIGHER_ORDER_MODEL}
\title{HIGHER_ORDER_MODEL class environment}
\format{
An environment with S3 class \code{oop}.
}
\usage{
HIGHER_ORDER_MODEL
}
\description{
This is the class of visual inference higher order linear model,
inherited from \link{VI_MODEL}.
}
\seealso{
Parent class: \link{VI_MODEL}
\cr
\cr
New attributes: \link{HIGHER_ORDER_MODEL$formula},
\link{HIGHER_ORDER_MODEL$null_formula}, \link{HIGHER_ORDER_MODEL$alt_formula},
\cr
\cr
New methods: \link{HIGHER_ORDER_MODEL$..init..}
}
\keyword{datasets}
