% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{is_instance}
\alias{is_instance}
\title{Check whether an environment is an instance built by a class constructor}
\usage{
is_instance(env, cls)
}
\arguments{
\item{env}{Environment. The instance environment.}

\item{cls}{Function. The class constructor.}
}
\value{
True or False.
}
\description{
This function returns \code{True} if the environment is an instance built by a
class constructor by checking the attribute "pseudo_oop_class"
of the class constructor.
}
\details{
This function only works if the class constructor is registered.
}
\examples{

tt <- BASE()
is_instance(tt, BASE)

}
