% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{is_subclass}
\alias{is_subclass}
\title{Check whether a class is a subclass of another class}
\usage{
is_subclass(child_cls, parent_cls)
}
\arguments{
\item{child_cls}{Function. Child class constructor.}

\item{parent_cls}{Function. Parent class constructor.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
This function returns \code{TRUE} if the class is a subclass of another class by
checking the attribute "pseduo_oop_class" of the class constructors.
}
\details{
This function only works if both class constructors are registered.
}
\examples{

# Define a derived class constructor
MY_CLASS <- function(..., env = new.env(parent = parent.frame())) {

  # Declare inheritance
  env <- inherit(env, BASE, "MY_CLASS", ...)

  # Methods of this derived class
  string_ <- function() "(MY_CLASS object)"

  # Register the method
  register_method(env, string = string_)

  # Return the instance environment
  return(env)
}

# Register the class constructor
register_class_ctor(MY_CLASS, "MY_CLASS", parent = BASE)

is_subclass(MY_CLASS, BASE)

}
\seealso{
\code{\link[=register_class_ctor]{register_class_ctor()}}
}
