% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{linear_approx_dirichlet}
\alias{linear_approx_dirichlet}
\title{Use linear approximation to calculate p-value for a visual test}
\usage{
linear_approx_dirichlet(c_i, n_eval, n_plot = 20, alpha = 1)
}
\arguments{
\item{c_i}{Numeric. Sum of the weighted detections.}

\item{n_eval.}{Number of evaluations.}

\item{n_plot.}{Number of plots.}
}
\value{
A named vector representing the probability mass function of the distribution.
}
\description{
This functions use linear approximation to calculate p-value for
a visual test. The Dirichlet attractiveness distribution
will be used.
}
\examples{

# Two people detect the data plot with one selection and one person
# detects the data plot with two selections.
linear_approx_dirichlet(3.5, 5)
}
