% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{new_class}
\alias{new_class}
\title{Define a new class}
\usage{
new_class(env = new.env(parent = parent.frame()), ..., class_name = NULL)
}
\arguments{
\item{env}{Environment. The new class environment.}

\item{...}{Environments. Parent class environments.}

\item{class_name}{Name of the new class.}
}
\value{
A class environment with S3 class "oop".
}
\description{
This function declare a new class, and copies attributes and methods from
parent classes.
}
\details{
Parents can be provided in \code{...}, where methods and attributes will be
overrided by the left classes.
}
\examples{
MYCLASS <- new_class(class_name = "MYCLASS")
MYCLASS
names(MYCLASS)

}
