% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{NON_NORMAL_MODEL$..init..}
\alias{NON_NORMAL_MODEL$..init..}
\title{Initialization method}
\arguments{
\item{x}{Random variable or closed form expression. Default is
\code{x = rand_uniform(-1, 1, env = new.env(parent = parent.env(self)))}.}

\item{e}{Random variable or closed form expression. Default is
\code{e = rand_lognormal(0, sigma, env = new.env(parent = parent.env(self)))}.}
}
\value{
Return the object itself.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment. The response variable of this model
is \code{y}. The formula of y is defined in \link{NON_NORMAL_MODEL$formula}, the null
formula is defined in \link{NON_NORMAL_MODEL$null_formula}, the alternative is
defined in \link{NON_NORMAL_MODEL$alt_formula}.
}
\examples{

# Instantiate
x <- rand_uniform()
e <- rand_lognormal(sigma = 0.5)

test <- non_normal_model(x = x, e = e)

test

# Generate data
test$gen(10)

# Generate lineup
test$gen_lineup(10, k = 3)

# Plot the lineup
test$plot_lineup(test$gen_lineup(100))

test <- non_normal_model(x = x, e = rand_lognormal(sigma = 0.1))
test$plot_lineup(test$gen_lineup(100))

test <- non_normal_model(x = x, e = rand_lognormal(sigma = 0.5))
test$plot_lineup(test$gen_lineup(100))

test <- non_normal_model(x = x, e = rand_lognormal(sigma = 1))
test$plot_lineup(test$gen_lineup(100))

test <- non_normal_model(x = x, e = rand_lognormal(sigma = 2))
test$plot_lineup(test$gen_lineup(100))
}
