% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\docType{data}
\name{NON_NORMAL_MODEL}
\alias{NON_NORMAL_MODEL}
\title{NON_NORMAL_MODEL class environment}
\format{
An environment with S3 class \code{bandicoot_oop}.
}
\usage{
NON_NORMAL_MODEL
}
\description{
This is the class of visual inference orthogonal polynomial
linear model, inherited from \link{VI_MODEL}.
}
\seealso{
Parent class: \link{VI_MODEL}
\cr
\cr
New attributes: \link{NON_NORMAL_MODEL$formula},
\link{NON_NORMAL_MODEL$null_formula}, \link{NON_NORMAL_MODEL$alt_formula}
\cr
\cr
New methods: \link{NON_NORMAL_MODEL$..init..}, \link{NON_NORMAL_MODEL$test}
}
\keyword{datasets}
