% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{POLY_MODEL$..init..}
\alias{POLY_MODEL$..init..}
\title{Initialization method}
\arguments{
\item{a}{Integer. The degree of the orthogonal polynomial used in the model.
Note that only the \code{a} degree term will be used, and it should be a value
between 2 to 6. Default is \code{a = 2}.}

\item{sigma}{Positive numeric. Default is \code{sigma = 1}.}

\item{x}{Random variable or closed form expression. Default is
\code{x = rand_uniform(-1, 1, env = new.env(parent = parent.env(self)))}.}

\item{e}{Random variable or closed form expression. Default is
\code{e = rand_normal(0, sigma, env = new.env(parent = parent.env(self)))}.}
}
\value{
No return value, called for side effects.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment. The response variable of this model
is \code{y}. The formula of y is defined in \link{POLY_MODEL$formula}, the null
formula is defined in \link{POLY_MODEL$null_formula}, the alternative is
defined in \link{POLY_MODEL$alt_formula}. The formula for the raw
orthogonal polynomial term is defined in \link{POLY_MODEL$raw_z_formula}, and
the scaled orthogonal polynomial term is defined in \link{POLY_MODEL$z_formula}.
}
\examples{

# Instantiation
x <- rand_uniform()
e <- rand_normal(sigma = 0.5)

test <- poly_model(a = 4, x = x, e = e)

test

# Generate data
test$gen(10)

# Generate lineup
test$gen_lineup(10, k = 3)

# Plot the lineup
test$plot_lineup(test$gen_lineup(100))
}
