% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{POLY_MODEL$E}
\alias{POLY_MODEL$E}
\title{Expectation of the residuals}
\arguments{
\item{dat}{Dataframe/List. List contains variable \code{x} and \code{z}.}
}
\value{
A vector of numeric expectations.
}
\description{
This function calculate the expectation of the residuals by the
use of the Frisch–Waugh–Lovell theorem.
}
\examples{

mod <- poly_model(4, 0.5)
dat <- mod$gen(1000, fit_model = TRUE)
dat$exp <- mod$E(dat)
mod$plot(dat) + ggplot2::geom_point(ggplot2::aes(.fitted, exp),
                                    col = "red",
                                    alpha = 0.6)
}
