% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{POLY_MODEL$effect_size}
\alias{POLY_MODEL$effect_size}
\title{Compute the effect size of the simulated data}
\arguments{
\item{dat}{Dataframe/List. List contains variable \code{x} and \code{z}.}

\item{sigma}{Positive numeric. Default is \code{sigma = self$prm$sigma}.}
}
\value{
A numeric value.
}
\description{
This function computes the effect size of the simulated data.
}
\examples{

mod <- poly_model(4, 0.5)
dat <- mod$gen(1000, fit_model = TRUE)
mod$effect_size(dat)
}
