% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{POLY_MODEL$hermite}
\alias{POLY_MODEL$hermite}
\title{Hermite polynomial functions}
\arguments{
\item{shape}{Integer. A value between 1 to 4.}
}
\value{
A function.
}
\description{
This function returns a Hermite polynomial function defined
using \link[mpoly:hermite]{mpoly::hermite}.
}
\examples{

POLY_MODEL$hermite(1)
POLY_MODEL$hermite(2)
POLY_MODEL$hermite(3)
POLY_MODEL$hermite(4)
}
