% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{POLY_MODEL$raw_z_formula}
\alias{POLY_MODEL$raw_z_formula}
\title{Formula for the raw orthogonal polynomial term \code{raw_z}}
\description{
Quoted formula for the raw orthogonal polynomial term \code{raw_z}.
}
\examples{

POLY_MODEL$z_formula
}
