% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{POLY_MODEL$sample_effect_size}
\alias{POLY_MODEL$sample_effect_size}
\title{Compute the sample based effect size of the simulated data}
\arguments{
\item{dat}{Dataframe/List. List contains variable \code{x} and \code{z}.}

\item{sigma}{Positive numeric. Default is \code{sigma = self$prm$sigma}.}

\item{include_z}{Boolean. Default is \code{include_z = self$prm$include_z}.}

\item{shape}{Positive integer. Default is \code{shape = self$prm$shape}.}

\item{type}{Character. Type of the effect size measure. Default is
\code{type = kl}.}
}
\value{
A numeric value.
}
\description{
This function computes the sample based effect size of the simulated data.
}
\examples{

mod <- poly_model(4, 0.5)
dat <- mod$gen(1000, fit_model = TRUE)
mod$sample_effect_size(dat)
}
