% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{POLY_MODEL$set_prm}
\alias{POLY_MODEL$set_prm}
\title{Set parameter for the model}
\arguments{
\item{prm_name}{List or Vector. Parameter character names.}

\item{prm_val}{List or Vector. Parameter values.}
}
\value{
Return the object itself.
}
\description{
This function store the values in the environment and update
their values in the closed form expression of \code{y}, except the parameter
\code{sigma}, \code{a} and \code{raw_z}. For parameter \code{sigma}, its value will be updated,
and the corresponding value in \code{e} will be updated. For parameter \code{a}, its
value will be updated, and the corresponding value in \code{raw_z} will be
updated. For parameter \code{raw_z}, its value will be updated, and the
corresponding value in \code{z} will be updated.
}
\examples{

# Instantiation
mod <- poly_model(a = 2, sigma = 0.5)

mod

mod$set_prm("a", 4)

mod

mod$set_prm("sigma", 1)

mod
}
