% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{POLY_MODEL$test}
\alias{POLY_MODEL$test}
\title{Test the null model}
\arguments{
\item{dat}{Data frame. A data frame containing all variables needed by the
\code{null_formula}.}

\item{null_formula}{Formula. Formula for fitting the null model. Default
is \code{null_formula = self$null_formula}.}

\item{alt_formula}{Formula. Formula for fitting the alternative model. Only
used in F-test.}

\item{test}{Character. A string representing the test. Either "F" or "RESET".}

\item{power}{Integer. A vector of integers representing the power of the
variables that should be included. Only used in RESET test. Default is 2:3.}

\item{power_type}{Character. Argument passed to \link[lmtest:resettest]{lmtest::resettest}. "fitted",
"regressor" or "princomp". Only used in RESET test. Default is "fitted".}
}
\value{
A list containing the test name, the test statistic and the p-value.
}
\description{
This function test the null model.
In this class, the test is either a F-test computed using \link[stats:anova]{stats::anova} or a
RESET test computed using \link[lmtest:resettest]{lmtest::resettest}.
}
\examples{

# Instantiate
mod <- poly_model(shape = 2, sigma = 0.5)

mod

dat <- mod$gen(100000)

# F-test
mod$test(dat, test = "F")

# RESET test
mod$test(dat, test = "RESET", power = 2:4)
}
