% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{POLY_MODEL$z_formula}
\alias{POLY_MODEL$z_formula}
\title{Formula for the scaled orthogonal polynomial term \code{z}}
\description{
Quoted formula for the scaled orthogonal polynomial term \code{z}.
}
\examples{

POLY_MODEL$z_formula
}
