% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\docType{data}
\name{POLY_MODEL}
\alias{POLY_MODEL}
\title{POLY_MODEL class environment}
\format{
An environment with S3 class \code{visage_oop}.
}
\usage{
POLY_MODEL
}
\description{
This is the class of visual inference orthogonal polynomial
linear model, inherited from \link{VI_MODEL}.
}
\seealso{
Parent class: \link{VI_MODEL}
\cr
\cr
New attributes: \link{POLY_MODEL$formula},
\link{POLY_MODEL$null_formula}, \link{POLY_MODEL$alt_formula}, \link{POLY_MODEL$z_formula}
\cr
\cr
New methods: \link{POLY_MODEL$..init..}, \link{POLY_MODEL$E},
\link{POLY_MODEL$effect_size}, \link{POLY_MODEL$set_prm}
}
\keyword{datasets}
