% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{print.visage_oop}
\alias{print.visage_oop}
\title{S3 method of printing \code{visage_oop} object}
\usage{
\method{print}{visage_oop}(x, ...)
}
\arguments{
\item{x}{\code{visage_oop} object.}

\item{...}{ignored.}
}
\value{
No return value, called for side effects.
}
\description{
This function print the string representation of the object.
}
