% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{QUARTIC_MODEL$sample_effect_size}
\alias{QUARTIC_MODEL$sample_effect_size}
\title{Compute the sample based effect size of the simulated data}
\arguments{
\item{dat}{Dataframe/List. List contains variable \code{x}.}

\item{a}{Numeric. Default is \code{a = self$prm$a}.}

\item{b}{Numeric. Default is \code{b = self$prm$b}.}

\item{c}{Numeric. Default is \code{c = self$prm$c}.}

\item{sigma}{Positive numeric. Default is \code{sigma = self$prm$sigma}.}
}
\value{
A numeric value.
}
\description{
This function computes the sample based effect size of the simulated data.
}
\examples{

mod <- quartic_model(-1, 1, 1, 0.5)
dat <- mod$gen(1000, fit_model = TRUE)
mod$sample_effect_size(dat)
}
