% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{QUARTIC_MODEL$set_prm}
\alias{QUARTIC_MODEL$set_prm}
\title{Set parameter for the model}
\arguments{
\item{prm_name}{List or Vector. Parameter character names.}

\item{prm_val}{List or Vector. Parameter values.}
}
\value{
Return the object itself.
}
\description{
This function store the values in the environment and update
their values in the closed form expression of \code{y}, except the parameter
\code{sigma}. For parameter \code{sigma}, its value will be updated, and the
corresponding value in \code{e} will be updated.
}
\examples{

# Instantiate
mod <- quartic_model(a = -1, b = 1, c = 1, sigma = 0.5)

mod

mod$set_prm("a", 2)

mod

mod$set_prm("sigma", 1)

mod
}
