% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\docType{data}
\name{QUARTIC_MODEL}
\alias{QUARTIC_MODEL}
\title{QUARTIC_MODEL class environment}
\format{
An environment with S3 class \code{visage_oop}.
}
\usage{
QUARTIC_MODEL
}
\description{
This is the class of visual inference quartic linear model,
inherited from \link{VI_MODEL}.
}
\seealso{
Parent class: \link{VI_MODEL}
\cr
\cr
New attributes: \link{QUARTIC_MODEL$formula},
\link{QUARTIC_MODEL$null_formula}, \link{QUARTIC_MODEL$alt_formula},
\cr
\cr
New methods: \link{QUARTIC_MODEL$..init..}, \link{QUARTIC_MODEL$E},
\link{QUARTIC_MODEL$effect_size}
}
\keyword{datasets}
