% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_LOGNORMAL$..init..}
\alias{RAND_LOGNORMAL$..init..}
\title{Initialization method}
\arguments{
\item{mu}{Numeric. Mean of the log of the random variable. Default is 0.}

\item{sigma}{Numeric. Standard deviation of the log the random variable.
Default is 1.}
}
\value{
No return value, called for side effects.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment.
}
\examples{

# Instantiation
test <- RAND_LOGNORMAL$instantiation(mu = 1, sigma = 2)
test
}
