% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_LOGNORMAL}
\alias{RAND_LOGNORMAL}
\alias{rand_lognormal}
\title{RAND_LOGNORMAL class environment}
\usage{
rand_lognormal(
  mu = 0,
  sigma = 1,
  env = new.env(parent = parent.frame()),
  init_call = sys.call()
)
}
\arguments{
\item{mu}{Numeric. Mean of the log of the random variable.}

\item{sigma}{Numeric. Mean of the log of the random variable.}

\item{env}{Environment. The instance environment.}

\item{init_call}{Call. Contents of the \code{..init_call...} It is recommended to leave it
as default.}
}
\value{
An instance environment.
}
\description{
This is the class of the log-normal random variable, inherited from
\link{RAND_VAR}. It is an environment with S3 class \code{bandicoot_oop}.
}
\section{Functions}{
\itemize{
\item \code{rand_lognormal()}: Class constructor, same as \code{RAND_LOGNORMAL$instantiate()}.

}}
\section{Class information}{
\subsection{Parent classes}{
\itemize{
\item Direct:
\itemize{
\item \link{RAND_VAR}
}
\item Indirect:
\itemize{
\item \link[bandicoot:BASE]{bandicoot::BASE}
}
}
}

\subsection{New methods}{
\itemize{
\item E
\itemize{
\item \code{RAND_LOGNORMAL$E()}
}
\item G
\itemize{
\item \code{\link[=RAND_LOGNORMAL$gen]{RAND_LOGNORMAL$gen()}}
}
\item I
\itemize{
\item \code{\link[=RAND_LOGNORMAL$..init..]{RAND_LOGNORMAL$..init..()}}
}
\item V
\itemize{
\item \code{RAND_LOGNORMAL$Var()}
}
}
}
}

