% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_NORMAL$gen}
\alias{RAND_NORMAL$gen}
\title{Generate random values}
\arguments{
\item{n}{Integer. Number of observations.}

\item{...}{Ignored.}
}
\value{
A vector of numeric values.
}
\description{
This function generates random values from the random variable.
}
\examples{

test <- RAND_NORMAL$instantiation(mu = 1, sigma = 2)
test$gen(10)
}
\seealso{
\code{\link[stats:Normal]{stats::rnorm()}}
}
