% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_T$..init..}
\alias{RAND_T$..init..}
\title{Initialization method}
\arguments{
\item{mu}{Numeric. Mean. Default is 0.}

\item{tau}{Numeric. Scale parameter. Default is 1.}

\item{df}{Integer. Degree of freedom. Default is 10.}
}
\value{
Return the object itself.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment.
}
\examples{

# Instantiate
test <- rand_t(mu = 0, tau = 1, df = 3)
test
}
