% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_T}
\alias{RAND_T}
\alias{rand_t}
\title{RAND_T class environment}
\usage{
rand_t(
  mu = 0,
  tau = 1,
  df = 10,
  env = new.env(parent = parent.frame()),
  init_call = sys.call()
)
}
\arguments{
\item{mu}{Numeric.}

\item{tau}{Numeric. Scale parameter.}

\item{df}{Integer. Degree of freedom.}

\item{env}{Environment. The instance environment.}

\item{init_call}{Call. Contents of the \code{..init_call...} It is recommended to leave it
as default.}
}
\value{
An instance environment.
}
\description{
This is the class of the student-t random variable, inherited from
\link{RAND_VAR}. It is an environment with S3 class \code{bandicoot_oop}.
}
\section{Functions}{
\itemize{
\item \code{rand_t()}: Class constructor, same as \code{RAND_T$instantiate()}.

}}
\section{Class information}{
\subsection{Parent classes}{
\itemize{
\item Direct:
\itemize{
\item \link{RAND_VAR}
}
\item Indirect:
\itemize{
\item \link[bandicoot:BASE]{bandicoot::BASE}
}
}
}

\subsection{New methods}{
\itemize{
\item E
\itemize{
\item \code{RAND_T$E()}
}
\item G
\itemize{
\item \code{\link[=RAND_T$gen]{RAND_T$gen()}}
}
\item I
\itemize{
\item \code{\link[=RAND_T$..init..]{RAND_T$..init..()}}
}
\item V
\itemize{
\item \code{RAND_T$Var()}
}
}
}
}

