% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_UNIFORM$..init..}
\alias{RAND_UNIFORM$..init..}
\title{Initialization method}
\arguments{
\item{a}{Numeric. Lower bound.}

\item{b}{Numeric. Upper bound.}
}
\value{
Return the object itself.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{RAND_UNIFORM$..init..(a = 0, b = 1)
}\if{html}{\out{</div>}}
}
}
\examples{

# Instantiate
test <- rand_uniform(a = 1, b = 2)
test
}
