% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_UNIFORM}
\alias{RAND_UNIFORM}
\alias{rand_uniform}
\title{RAND_UNIFORM class environment}
\usage{
rand_uniform(
  a = 0,
  b = 1,
  env = new.env(parent = parent.frame()),
  init_call = sys.call()
)
}
\arguments{
\item{a}{Numeric. Lower bound.}

\item{b}{Numeric. Upper bound.}

\item{env}{Environment. The instance environment.}

\item{init_call}{Call. Contents of the \code{..init_call...} It is recommended to leave it
as default.}
}
\value{
An instance environment.
}
\description{
This is the class of the uniform random variable, inherited from
\link{RAND_VAR}. It is an environment with S3 class \code{bandicoot_oop}.
}
\section{Functions}{
\itemize{
\item \code{rand_uniform()}: Class constructor, same as \code{RAND_UNIFORM$instantiate()}.

}}
\section{Class information}{
\subsection{Parent classes}{
\itemize{
\item Direct:
\itemize{
\item \link{RAND_VAR}
}
\item Indirect:
\itemize{
\item \link[bandicoot:BASE]{bandicoot::BASE}
}
}
}

\subsection{New methods}{
\itemize{
\item E
\itemize{
\item \code{RAND_UNIFORM$E()}
}
\item G
\itemize{
\item \code{\link[=RAND_UNIFORM$gen]{RAND_UNIFORM$gen()}}
}
\item I
\itemize{
\item \code{\link[=RAND_UNIFORM$..init..]{RAND_UNIFORM$..init..()}}
}
\item V
\itemize{
\item \code{RAND_UNIFORM$Var()}
}
}
}
}

