% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_UNIFORM_D$..init..}
\alias{RAND_UNIFORM_D$..init..}
\title{Initialization method}
\arguments{
\item{a}{Numeric. Lower bound.}

\item{b}{Numeric. Upper bound.}

\item{k}{Integer. Number of unique discrete values.}

\item{even}{Boolean. Whether or not candidate values are evenly spaced.}
}
\value{
Return the object itself.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{RAND_UNIFORM_D$..init..(a = 0, b = 1, k = 5, even = FALSE)
}\if{html}{\out{</div>}}
}
}
\examples{

# Instantiate
test <- rand_uniform_d(a = 1, b = 2, k = 3)
test
}
