% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_UNIFORM_D}
\alias{RAND_UNIFORM_D}
\alias{rand_uniform_d}
\title{RAND_UNIFORM_D class environment}
\usage{
rand_uniform_d(
  a = 0,
  b = 1,
  k = 5,
  even = FALSE,
  env = new.env(parent = parent.frame()),
  init_call = sys.call()
)
}
\arguments{
\item{a}{Numeric. Lower bound.}

\item{b}{Numeric. Upper bound.}

\item{k}{Integer. Number of unique discrete values.}

\item{even}{Boolean. Whether or not candidate values are evenly spaced.}

\item{env}{Environment. The instance environment.}

\item{init_call}{Call. Contents of the \code{..init_call...} It is recommended to leave it
as default.}
}
\value{
An instance environment.
}
\description{
This is the class of the discrete uniform random variable, inherited from
\link{RAND_VAR}. It is an environment with S3 class \code{bandicoot_oop}.
}
\section{Functions}{
\itemize{
\item \code{rand_uniform_d()}: Class constructor, same as \code{RAND_UNIFORM_D$instantiate()}.

}}
\section{Class information}{
\subsection{Parent classes}{
\itemize{
\item Direct:
\itemize{
\item \link{RAND_VAR}
}
\item Indirect:
\itemize{
\item \link[bandicoot:BASE]{bandicoot::BASE}
}
}
}

\subsection{New methods}{
\itemize{
\item E
\itemize{
\item \code{RAND_UNIFORM_D$E()}
}
\item G
\itemize{
\item \code{\link[=RAND_UNIFORM_D$gen]{RAND_UNIFORM_D$gen()}}
}
\item I
\itemize{
\item \code{\link[=RAND_UNIFORM_D$..init..]{RAND_UNIFORM_D$..init..()}}
}
\item V
\itemize{
\item \code{RAND_UNIFORM_D$Var()}
}
}
}
}

