% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_VAR}
\alias{RAND_VAR}
\alias{rand_var}
\title{RAND_VAR class environment}
\usage{
rand_var(
  dist = "",
  prm = list(),
  env = new.env(parent = parent.frame()),
  init_call = sys.call()
)
}
\arguments{
\item{dist}{Character. Distribution name.}

\item{prm}{List. List of parameters.}

\item{env}{Environment. The instance environment.}

\item{init_call}{Call. Contents of the \code{..init_call...} It is recommended to leave it
as default.}
}
\value{
An instance environment.
}
\description{
This is the base class of random variable, inherited from
\link[bandicoot:BASE]{bandicoot::BASE}. It is an environment with S3 class \code{bandicoot_oop}.
}
\section{Functions}{
\itemize{
\item \code{rand_var()}: Class constructor, same as \code{RAND_VAR$instantiate()}.

}}
\section{Class information}{
\subsection{Parent classes}{
\itemize{
\item Direct:
\itemize{
\item \link[bandicoot:BASE]{bandicoot::BASE}
}
}
}

\subsection{New attributes}{
\itemize{
\item D:
\itemize{
\item \link{RAND_VAR$dist}
}
\item P:
\itemize{
\item \link{RAND_VAR$prm}
}
}
}

\subsection{New methods}{
\itemize{
\item E:
\itemize{
\item \code{\link[=RAND_VAR$E]{RAND_VAR$E()}}
}
\item I:
\itemize{
\item \code{\link[=RAND_VAR$..init..]{RAND_VAR$..init..()}}
}
\item S:
\itemize{
\item \code{\link[=RAND_VAR$set_prm]{RAND_VAR$set_prm()}}
\item \code{\link[=RAND_VAR$..str..]{RAND_VAR$..str..()}}
}
\item V:
\itemize{
\item \code{\link[=RAND_VAR$Var]{RAND_VAR$Var()}}
}
}
}
}

