% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{register_class_ctor}
\alias{register_class_ctor}
\title{Register class constructor}
\usage{
register_class_ctor(cls, cls_name, parent = NULL, ...)
}
\arguments{
\item{cls}{Function. The class constructor.}

\item{cls_name}{Character. Class name of the class constructor.}

\item{parent}{Function. The parent class constructor.}

\item{...}{ignored.}
}
\value{
No return value, called for side effects.
}
\description{
This function save the class information in the \code{pseudo_oop_class} attribute
of the class constructor.
}
\details{
This function needs to be called followed by the definition of the class
constructor.
}
\examples{

# Define a derived class
myclass <- function(..., env = new.env(parent = parent.frame())) {
  env <- inherit(env, BASE, "myclass", ...)

  myfunc_ <- function() 1 + 1

  register_method(env, myfunc = myfunc_)
  return(env)
}

register_class_ctor(myclass, "myclass", parent = BASE)
}
