% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{register_method}
\alias{register_method}
\title{Register method for a class or an instance}
\usage{
register_method(env, ...)
}
\arguments{
\item{env}{Environment. Class or instance environment.}

\item{...}{Named Functions. Functions needs to be provided in named format
like \code{a = function() 1}.}
}
\value{
No return value, called for side effects.
}
\description{
This function register a function as method of a class or an instance.
}
\details{
Method needs to be provided as \code{a = function() 1}, where \code{a} is the name of
the method and the right hand side of the equal sign is the function.
}
\examples{

x <- 0
a <- function() print(x)
a()

e <- new.env()
e$x <- 1
register_method(e, aa = a)

e$aa()

}
