% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_visage}
\alias{run_visage}
\title{Launch the ViSAGE R Shiny app}
\usage{
run_visage()
}
\value{
A Shiny app object that launches the ViSAGE application.
}
\description{
The `run_visage()` function launches the ViSAGE R Shiny app.
The app is designed to simulate and visualize various genetic
architectures and selection schemes in breeding populations.
The app provides an interactive interface for users to
explore and compare different selection strategies.
It also allows the user to run GWAS and genomic prediction analyses using either
simulated data or user-inputed data.
}
\examples{
# To run the ViSAGE app, simply call the function as shown below:
run_visage()
}
