% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{SIMPLE_CUBIC_MODEL$..init..}
\alias{SIMPLE_CUBIC_MODEL$..init..}
\title{Initialization method}
\arguments{
\item{a}{Numeric. Default is \code{a = 1}.}

\item{b}{Numeric. Default is \code{b = 1}.}

\item{sigma}{Positive numeric. Default is \code{sigma = 1}.}

\item{x}{Random variable or closed form expression. Default is
\code{x = rand_uniform(-1, 1, env = new.env(parent = parent.env(self)))}.}

\item{e}{Random variable or closed form expression. Default is
\code{e = rand_normal(0, sigma, env = new.env(parent = parent.env(self)))}.}
}
\value{
No return value, called for side effects.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment. The response variable of this model
is \code{y}. The formula of y is defined in \link{SIMPLE_CUBIC_MODEL$formula}, the null
formula is defined in \link{SIMPLE_CUBIC_MODEL$null_formula}, the alternative is
defined in \link{SIMPLE_CUBIC_MODEL$alt_formula}.
}
\examples{

# Instantiation
x <- rand_uniform()
e <- rand_normal()

test <- simple_cubic_model(a = 200, b = 200, x = x, e = e)

test

# Generate data
test$gen(10)

# Generate lineup
test$gen_lineup(10, k = 3)

# Plot the lineup
test$plot_lineup(test$gen_lineup(100))
}
