% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{SIMPLE_CUBIC_MODEL$null_formula}
\alias{SIMPLE_CUBIC_MODEL$null_formula}
\title{Formula for fitting the null model}
\description{
Quoted formula for fitting the null model.
}
\examples{

SIMPLE_CUBIC_MODEL$null_formula
}
