% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\docType{data}
\name{SIMPLE_CUBIC_MODEL}
\alias{SIMPLE_CUBIC_MODEL}
\title{SIMPLE_CUBIC_MODEL class environment}
\format{
An environment with S3 class \code{bandicoot_oop}.
}
\usage{
SIMPLE_CUBIC_MODEL
}
\description{
This is the class of visual inference simple cubic linear model,
inherited from \link{VI_MODEL}.
}
\seealso{
Parent class: \link{VI_MODEL}
\cr
\cr
New attributes: \link{SIMPLE_CUBIC_MODEL$formula},
\link{SIMPLE_CUBIC_MODEL$null_formula}, \link{SIMPLE_CUBIC_MODEL$alt_formula},
\cr
\cr
New methods: \link{SIMPLE_CUBIC_MODEL$..init..}, \link{SIMPLE_CUBIC_MODEL$E},
\link{SIMPLE_CUBIC_MODEL$sample_effect_size}, \link{SIMPLE_CUBIC_MODEL$set_prm}
}
\keyword{datasets}
