% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$..cache..}
\alias{VI_MODEL$..cache..}
\title{Cache list, containing the last fitted model, data frame and formula}
\description{
A list, will be used if \code{cache = TRUE} while calling the \code{fit}
method.
}
\examples{

# Instantiation
x <- RAND_UNIFORM$instantiation()
test <- VI_MODEL$instantiation(prm = list(x = x),
                               prm_type = list(x = "r"),
                               formula = y~x+x^2,
                               null_formula = y~x,
                               alt_formula = y~x+x^2)

# Turn on cache
test$fit(test$gen(10), cache = TRUE)

test$..cache..
}
