% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$..str..}
\alias{VI_MODEL$..str..}
\title{String representation of the object}
\value{
A string.
}
\description{
This function returns a string representation of the object.
}
\examples{

VI_MODEL$..str..()

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y~x+x^2+e,
                 null_formula = y~x,
                 alt_formula = y~x+I(x^2))

test$..str..()
}
