% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$alt_formula}
\alias{VI_MODEL$alt_formula}
\title{Formula for fitting the alternative model}
\description{
A formula, will be used to fit the alternative model.
}
\examples{

# Instantiation
x <- RAND_UNIFORM$instantiation()
test <- VI_MODEL$instantiation(prm = list(x = x),
                               prm_type = list(x = "r"),
                               formula = y~x+x^2,
                               null_formula = y~x,
                               alt_formula = y~x+I(x^2))

# A formula with environment equals to `test`
test$alt_formula

# Use alternative formula in `fit`
test$fit(test$gen(10), test$alt_formula)

# F-test also needs to use the alternative model
test$test(test$gen(10))
}
