% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$effect_size}
\alias{VI_MODEL$effect_size}
\title{Compute the effect size of the simulated data or the defined model}
\arguments{
\item{...}{Derived class needs to override this method.}
}
\value{
NA
}
\description{
This function computes the effect size of the simulated data or
the defined model. This is an abstract method.
}
\examples{

VI_MODEL$effect_size()
}
