% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$fit}
\alias{VI_MODEL$fit}
\title{Test the null model against the alternative model}
\arguments{
\item{dat}{Data frame. A data frame containing all variables needed by the
\code{formula}.}

\item{formula}{Formula. Formula for fitting the model. Default
is \code{formula = self$null_formula}.}

\item{...}{Arguments passed to \code{stats::lm}.}
}
\value{
A fitted linear model.
}
\description{
This function fit a linear model by using \code{stats::lm}.
}
\examples{

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$fit(test$gen(100000))

test$fit(test$gen(100000), formula = y ~ x + I(x^2))
}
