% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$formula}
\alias{VI_MODEL$formula}
\title{Closed form expression of \code{y}}
\description{
A formula, will be passed to \code{CLOSED_FORM$instantiation} to
define a closed form expression for \code{y}.
}
\examples{

# Instantiation
x <- RAND_UNIFORM$instantiation()
test <- VI_MODEL$instantiation(prm = list(x = x),
                               prm_type = list(x = "r"),
                               formula = y~x+x^2,
                               null_formula = y~x,
                               alt_formula = y~x+I(x^2))

# A formula with environment equals to `test`
test$formula

# `y` has the same expression
test$prm$y
}
