% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$gen}
\alias{VI_MODEL$gen}
\title{Generating random values from the expression of \code{y}}
\arguments{
\item{n}{Integer. Number of observations.}

\item{fit_model}{Boolean. Whether or not to fit a null model to obtain the
fitted values and the residuals. Default is \code{TRUE}.}

\item{test}{Boolean. Whether or not to test the null model against the
alternative model to obtain the test statistic and the p-value.
Default is \code{FALSE}.}

\item{computed}{List. Default is \code{NULL}. If it is provided, random variables
or random closed form expression will use the values from the list, which
makes the expression potentially deterministic.}
}
\value{
A data frame.
}
\description{
This function generates random values from the expression of
\code{y}, and keeps all the right hand side information in a data frame.
}
\examples{

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

dat <- test$gen(10)
dat

test$gen(10, computed = list(e = dat$e))

test$gen(10, test = TRUE)
}
