% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$gen_lineup}
\alias{VI_MODEL$gen_lineup}
\title{Generating random values from the expression of \code{y}, and forms a lineup}
\arguments{
\item{n}{Integer. Number of observations.}

\item{k}{Integer. Number of plots in the lineup. Default is \code{k = 20}.}

\item{pos}{Integer. Position of the true data plot. Default is \code{pos = NULL},
which means the position is random.}

\item{computed}{List. Default is \code{NULL}. If it is provided, random variables
or random closed form expression will use the values from the list, which
makes the expression potentially deterministic.}
}
\value{
A data frame.
}
\description{
This function generates random values from the expression of
\code{y}, and keeps all the right hand side information in a data frame.
}
\examples{

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$gen_lineup(10, k = 3)

test$gen_lineup(10, k = 3, computed = list(e = 1:10))
}
