% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$null_formula}
\alias{VI_MODEL$null_formula}
\title{Formula for fitting the null model}
\description{
A formula, will be used to fit the null model.
}
\examples{

# Instantiation
x <- RAND_UNIFORM$instantiation()
test <- VI_MODEL$instantiation(prm = list(x = x),
                               prm_type = list(x = "r"),
                               formula = y~x+x^2,
                               null_formula = y~x,
                               alt_formula = y~x+I(x^2))

# A formula with environment equals to `test`
test$null_formula

# By default, `fit` use the null formula
test$fit(test$gen(10))

# F-test also needs to use the null model
test$test(test$gen(10))
}
