% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$null_resid}
\alias{VI_MODEL$null_resid}
\title{Generate null residuals from a null model}
\arguments{
\item{dat}{Data frame. A data frame that used to fit the model.}

\item{mod}{Linear Model. The null model.}

\item{test}{Boolean. Whether or not to use \code{test} on the newly generated
data.}
}
\value{
A data frame with updated \code{y}, \code{.resid}, and potentially updated
test result.
}
\description{
This function uses the data and the null model to generate
null residuals. The null residuals are generated by first regressing random
noises on the original regressors, then multiply the obtained residuals by
original RSS divided by the current RSS. The results are the null residuals.
}
\examples{

# Instantiate
x <- bandicoot::rand_uniform()
e <- bandicoot::rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

dat <- test$gen(10)
test$null_resid(dat, test$fit(dat), test = TRUE)
}
