% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$plot}
\alias{VI_MODEL$plot}
\title{Plot the fitted model}
\arguments{
\item{dat}{Data frame. A data frame containing correspoding variables.}

\item{type}{Character. "resid": \link{VI_MODEL$plot_resid}. Default is
\code{type = "resid"}.}

\item{theme}{ggtheme. A ggplot theme object.}

\item{alpha}{Numeric. Alpha of dot. Value between 0 and 1. Default is 1.}

\item{size}{Numeric. Size of dot. Value between 0 and 1. Default is 0.5.}

\item{remove_axis}{Boolean. Whether or not to remove the axis. Default is
\code{remove_axis = FALSE}.}

\item{remove_legend}{Boolean. Whether or not to remove the legend. Default is
\code{remove_legend = FALSE}.}

\item{remove_grid_line}{Boolean. Whether or not to remove the grid lines.
Default is \code{remove_grid_line = FALSE}.}

\item{add_zero_line}{Boolean. Whether or not to add a zero horizontal line.
Default is \code{add_zero_line = TRUE}.}
}
\value{
A ggplot.
}
\description{
This function generates plot for fitted model.
}
\examples{

# Instantiate
x <- bandicoot::rand_uniform()
e <- bandicoot::rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + 10 * x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

# Residual plot of the null model
test$plot(test$gen(1000, fit_model = TRUE), type = "resid")

# Customize the plot
test$plot(test$gen(1000, fit_model = TRUE),
          type = "resid",
          theme = ggplot2::theme_light(),
          remove_axis = TRUE,
          remove_grid_line = TRUE,
          add_zero_line = TRUE)
}
