% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$plot_lineup}
\alias{VI_MODEL$plot_lineup}
\title{Plot the lineup}
\arguments{
\item{dat}{Data frame. A data frame containing correspoding variables.}

\item{type}{Character. "resid": \link{VI_MODEL$plot_resid}. Default is
\code{type = "resid"}.}

\item{...}{Arguments passed to \link{VI_MODEL$plot}}
}
\value{
A ggplot.
}
\description{
This function plots the lineup.
}
\examples{

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + 10 * x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

# Lineup of residual plots
test$plot_lineup(test$gen_lineup(100), type = "resid")

# Customize the plot
test$plot_lineup(test$gen_lineup(100),
                 type = "resid",
                 theme = ggplot2::theme_light(),
                 remove_axis = TRUE,
                 remove_grid_line = TRUE,
                 add_zero_line = TRUE)
}
