% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$plot_qq}
\alias{VI_MODEL$plot_qq}
\title{Plot the residual Q-Q plot}
\arguments{
\item{dat}{Data frame. A data frame containing \code{.resid}.}
}
\value{
A ggplot.
}
\description{
This function generate a Q-Q plot for residuals.
}
\examples{

# Instantiate
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + 10 * x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

# Residual plot of the null model
test$plot_qq(test$gen(1000, fit_model = TRUE))
}
