% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$prm_type}
\alias{VI_MODEL$prm_type}
\title{List of parameter types}
\description{
A list, will be initialized after an instance is built. "r"
stands for random variable or closed form used in the expression of \code{y},
"o" stands for others. This value only affects the string representation of
the object. If a variable is not part of the expression of \code{y}, then it
should be labelled as "o". If a variable is part of the expression of \code{y},
but it is not a random variable or a closed form expression, then it should
be labelled as "o" as well.
}
\examples{

# Instantiation
x <- RAND_UNIFORM$instantiation()
test <- VI_MODEL$instantiation(prm = list(x = x),
                               prm_type = list(x = "r"),
                               formula = y~x+x^2)
test$prm_type
}
