% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$rss}
\alias{VI_MODEL$rss}
\title{Residual sum of square of a fitted model}
\value{
Numeric value.
}
\description{
This function returns the residual sum of square of a fitted
model.
}
\examples{

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test$rss(test$fit(test$gen(100)))
}
