% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$sample_effect_size}
\alias{VI_MODEL$sample_effect_size}
\title{Compute the sample based effect size of the simulated data of the defined model}
\arguments{
\item{...}{Derived class needs to override this method.}
}
\value{
NA
}
\description{
This function computes the sample based effect size of the
simulated data of the defined model. This is an abstract method.
}
\examples{

VI_MODEL$sample_effect_size()
}
