% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$set_prm}
\alias{VI_MODEL$set_prm}
\title{Set parameter for the model}
\arguments{
\item{prm_name}{List or Vector. Parameter character names.}

\item{prm_val}{List or Vector. Parameter values.}
}
\value{
Return the object itself.
}
\description{
This function store the values in the environment and update
their values in the closed form expression of \code{y}. However, if the parameter
is not directly used by the closed form expression, user should override
this method to correctly set the parameter.
}
\examples{

# Instantiate
x <- bandicoot::rand_uniform()
e <- bandicoot::rand_normal()
test <- vi_model(prm = list(a = 1, x = x, e = e),
                 prm_type = list(a = "o", x = "r", e = "r"),
                 formula = y ~ 1 + a * x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

test

test$set_prm("a", 2)

test
}
