% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$test}
\alias{VI_MODEL$test}
\title{Test the null model against the alternative model}
\arguments{
\item{dat}{Data frame. A data frame containing all variables needed by the
\code{null_formula} and \code{alt_formula}.}

\item{null_formula}{Formula. Formula for fitting the null model. Default
is \code{null_formula = self$null_formula}.}

\item{alt_formula}{Formula. Formula for fitting the alternative model.
Default is \code{alt_formula = self$alt_formula}.}
}
\value{
A list containing the test name, the test statistic and the p-value.
}
\description{
This function test the null model against the alternative model.
In this class, the test is a F-test computed using \link[stats:anova]{stats::anova}. Derived
classes may have their own test procedures.
}
\examples{

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))


dat <- test$gen(100000)
test$test(dat)

test$test(dat, alt_formula = y ~ x + I(x^3))
}
