% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\docType{data}
\name{VI_MODEL}
\alias{VI_MODEL}
\title{VI_MODEL class environment}
\format{
An environment with S3 class \code{oop}.
}
\usage{
VI_MODEL
}
\description{
This is the base class of visual inference linear model,
inherited from \link{BASE}.
}
\seealso{
Parent class: \link{BASE}
\cr
\cr
New attributes: VI_MODEL$prm, VI_MODEL$prm_type, VI_MODEL$..cache..,
VI_MODEL$formula, VI_MODEL$null_formula, VI_MODEL$alt_formula
\cr
\cr
New methods: CLOSED_FORM$..init.., CLOSED_FORM$..str..,
CLOSED_FORM$compute, CLOSED_FORM$gen, CLOSED_FORM$ast
}
\keyword{datasets}
