% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{BASE}
\alias{BASE}
\title{Base class constructor}
\usage{
BASE(..., env = new.env(parent = parent.frame()))
}
\arguments{
\item{...}{Values that will be stored in the instance environment.}

\item{env}{Environment. The instance environment.}
}
\value{
An environment with S3 class "pseudo_oop".
}
\description{
This class includes some basic methods and attributes.
}
\details{
details
}
\examples{

base_instance <- BASE(name = "a")
base_instance
base_instance$methods()

}
