% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{BASE}
\alias{BASE}
\title{Base class constructor}
\usage{
BASE(..., env = new.env(parent = parent.frame()))
}
\arguments{
\item{...}{Values that will be stored in the instance environment.}

\item{env}{Environment. The instance environment.}
}
\value{
An environment with S3 class "pseudo_oop".
}
\description{
This class includes some basic methods and attributes.
}
\details{
The base class provides the following attributes and methods:
\itemize{
\item \code{class} : Class list.
\item \code{type} : Class.
\item \code{doc} : Doc string.
\item \code{init_call} : The call that init this instance.
\item \code{methods()} : Get all method names.
\item \code{has_attr(attr_name)} : Whether the instance has the attribute.
\item \code{set_attr(attr_name, attr_val)} : Set value for an attribute.
\item \code{get_attr(attr_name)} : Get the value of an attribute.
\item \code{del_attr(attr_name)} : Delete an attribute.
\item \code{list_attr()} : Get all attributes name.
\item \code{len()} : Length of the instance.
\item \code{repr()} : Representation of an object.
\item \code{string()} : String pepresentation of an object.
}
}
\examples{

base_instance <- BASE(name = "a")
base_instance
base_instance$methods()

}
