% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{calc_p_value}
\alias{calc_p_value}
\title{Calculate p-value for a visual test}
\usage{
calc_p_value(
  n_detections,
  n_evaluations,
  n_selections,
  n_plots = 20,
  n_simulations = 50000,
  cache_env = NULL,
  seed = NULL
)
}
\arguments{
\item{n_detections}{Integer. Observed value of number of detections.}

\item{n_evaluations, n_selections, n_plots, n_simulations}{Arguments passed to \code{\link[=sim_dist]{sim_dist()}}.}

\item{cache_env}{Environment. A provided environment for caching.}

\item{seed}{Integer. \code{\link[=set.seed]{set.seed()}} will be run at the beginning of the
function if \code{seed} is provided.}
}
\value{
A numeric value representing the p-value.
}
\description{
This function calculate the p-value for a visual test. The null distribution
is simulated by using function \code{\link[=sim_dist]{sim_dist()}}. The p-value is the probability
of the number of detections greater or equal to the observed value.
}
\details{
It is encouraged to provide a cache environment to boost up the performance
when this function needs to be reused. The cache environment will remember
the result corresponding to the combinations of \code{n_evaluations}
and \code{n_simulations}.
}
\examples{
calc_p_value(2, 3, c(1,1,2))
calc_p_value(1, 1, c(1))

}
