% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_CLOSED_FORM.R
\name{CLOSED_FORM$gen}
\alias{CLOSED_FORM$gen}
\title{Generating random values from the expression}
\arguments{
\item{n}{Integer. Number of observations.}

\item{rhs_val}{Boolean. Whether or not to keep the right hand side values of the
expression.}
}
\value{
Numeric values.
}
\description{
This function generates random values from the expression.
Random values will be generated independently as long as they don't share
the same symbol in the expression.
}
\examples{

# Constant variable
a <- 1

# Random uniform variable
b <- RAND_UNIFORM$instantiation()

# Define a closed form expression
cf <- CLOSED_FORM$instantiation(~3 * (exp(a) + b))

# Generate 5 values
cf$gen(5)

# Generate 5 values, and keep RHS
cf$gen(5, rhs_val = TRUE)

d <- RAND_NORMAL$instantiation()

# Define a closed form expression with another closed form expression
cf2 <- CLOSED_FORM$instantiation(~cf + 3 * d)

# Generate 5 values
cf2$gen(5)

# Generate 5 values, and keep RHS
cf2$gen(5, rhs_val = TRUE)

# Define a closed form expression with two random variables of the same name
cf3 <- CLOSED_FORM$instantiation(~d + d)

# Both `d` will share the same values
cf3$gen(5, rhs_val = TRUE)

# Define a closed form expression with two closed form expressions of the same name
cf4 <- CLOSED_FORM$instantiation(~cf3 + cf3)

# Both `cf3` will share the same values, both `d` will share the same values as well
cf4$gen(5, rhs_val = TRUE)

# Define a closed form expression with two different closed form expressions,
# but contains same random variables
cf5 <- CLOSED_FORM$instantiation(~cf3 + cf4)

# Both `d` in `cf3` and `cf4` will share the same value
cf5$gen(5, rhs_val = TRUE)
}
