% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_CLOSED_FORM.R
\docType{data}
\name{CLOSED_FORM}
\alias{CLOSED_FORM}
\title{CLOSED_FORM class environment}
\format{
An environment with S3 class \code{oop}.
}
\usage{
CLOSED_FORM
}
\description{
This is the base class of closed form expression, inherited from
\link{BASE}.
}
\seealso{
Parent class: \link{BASE}
\cr
\cr
New attributes: CLOSED_FORM$sym, CLOSED_FORM$sym_name,
CLOSED_FORM$type, CLOSED_FORM$expr
\cr
\cr
New methods: CLOSED_FORM$..init.., CLOSED_FORM$..str..,
CLOSED_FORM$compute, CLOSED_FORM$gen, CLOSED_FORM$gen_rhs,
CLOSED_FORM$ast
}
\keyword{datasets}
