% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{HETER_MODEL$..init..}
\alias{HETER_MODEL$..init..}
\title{Initialization method}
\arguments{
\item{a}{Numeric. Default is \code{a = 0}.}

\item{b}{Numeric. Default is \code{b = 1}.}

\item{x}{Random variable or closed form expression. Default is
\code{x = rand_uniform(-1, 1, env = new.env(parent = parent.env(self)))}.}
}
\value{
No return value, called for side effects.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment. The response variable of this model
is \code{y}. The formula of y is defined in \link{HETER_MODEL$formula},
the null formula is defined in \link{HETER_MODEL$null_formula},
the formula for \code{sigma} is defined in \link{HETER_MODEL$sigma_formula},
the formula for \code{e} is defined in \link{HETER_MODEL$e_formula}.
}
\examples{

# Instantiation
test <- heter_model(a = 0, b = 20)

test

# Generate data
test$gen(10)

# Generate lineup
test$gen_lineup(10, k = 3)

# Plot the lineup
test$plot_lineup(test$gen_lineup(100))
}
