% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\docType{data}
\name{HETER_MODEL}
\alias{HETER_MODEL}
\title{HETER_MODEL class environment}
\format{
An environment with S3 class \code{oop}.
}
\usage{
HETER_MODEL
}
\description{
This is the class of visual inference heteroskedasticity
linear model, inherited from \link{VI_MODEL}.
}
\seealso{
Parent class: \link{VI_MODEL}
\cr
\cr
New attributes: \link{HETER_MODEL$formula},
\link{HETER_MODEL$null_formula},
\link{HETER_MODEL$alt_formula},
\link{HETER_MODEL$sigma_quote_formula},
\link{HETER_MODEL$e_quote_formula}
\cr
\cr
New methods: \link{VI_MODEL$..init..}, \link{VI_MODEL$test}
}
\keyword{datasets}
