% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{inherit}
\alias{inherit}
\title{Declare inheritance inside a class constructor}
\usage{
inherit(env, parent, child_name, ...)
}
\arguments{
\item{env}{Environment. The instance environment.}

\item{parent}{Function. The parent class constructor.}

\item{child_name}{Character. The derived class name.}

\item{...}{Arguments passed to the parent class constructor.}
}
\value{
A parent class instance with updated class information.
}
\description{
This function builds an instance of the parent class, with updated
class information.
}
\details{
This function is assumed to be run inside a class constructor directly.
Otherwise, user needs to reset the \code{init_call} attribute by
calling \code{env$init_call <- sys.call()} inside the constructor directly.
}
\examples{

# Define a derived class constructor
MY_CLASS <- function(..., env = new.env(parent = parent.frame())) {

  # Declare inheritance
  env <- inherit(env, BASE, "MY_CLASS", ...)

  # Methods of this derived class
  myfunc_ <- function() 1 + 1

  # Register the method
  register_method(env, myfunc = myfunc_)

  # Return the instance environment
  return(env)
}

# Register the class constructor
register_class_ctor(MY_CLASS, "MY_CLASS", parent = BASE)

# Init an instance
my_instance <- MY_CLASS()

# Some methods are inherited from `BASE`
my_instance$methods()

}
