% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_zzz.R
\name{portal}
\alias{portal}
\alias{closed_form}
\alias{vi_model}
\alias{cubic_model}
\alias{simple_cubic_model}
\alias{quartic_model}
\alias{poly_model}
\alias{heter_model}
\alias{ar1_model}
\alias{non_normal_model}
\alias{phn_model}
\title{Portals to class instantiate methods}
\usage{
closed_form(
  ...,
  env = new.env(parent = parent.frame()),
  init_call = sys.call()
)

vi_model(..., env = new.env(parent = parent.frame()), init_call = sys.call())

cubic_model(
  ...,
  env = new.env(parent = parent.frame()),
  init_call = sys.call()
)

simple_cubic_model(
  ...,
  env = new.env(parent = parent.frame()),
  init_call = sys.call()
)

quartic_model(
  ...,
  env = new.env(parent = parent.frame()),
  init_call = sys.call()
)

poly_model(..., env = new.env(parent = parent.frame()), init_call = sys.call())

heter_model(
  ...,
  env = new.env(parent = parent.frame()),
  init_call = sys.call()
)

ar1_model(..., env = new.env(parent = parent.frame()), init_call = sys.call())

non_normal_model(
  ...,
  env = new.env(parent = parent.frame()),
  init_call = sys.call()
)

phn_model(..., env = new.env(parent = parent.frame()), init_call = sys.call())
}
\arguments{
\item{...}{Arguments passed to \code{init} method.}

\item{env}{Environment. The instance environment.}

\item{init_call}{Call. Contents of the \code{..init_call..}. It is recommended to
leave it as default. Default is \code{init_call = sys.call()}.}
}
\description{
These functions are the same as \code{class$instantiate}.
}
\section{Functions}{
\itemize{
\item \code{closed_form()}: \link{CLOSED_FORM}

\item \code{vi_model()}: \link{VI_MODEL}

\item \code{cubic_model()}: \link{CUBIC_MODEL}

\item \code{simple_cubic_model()}: \link{SIMPLE_CUBIC_MODEL}

\item \code{quartic_model()}: \link{QUARTIC_MODEL}

\item \code{poly_model()}: \link{POLY_MODEL}

\item \code{heter_model()}: \link{HETER_MODEL}

\item \code{ar1_model()}: \link{AR1_MODEL}

\item \code{non_normal_model()}: \link{NON_NORMAL_MODEL}

\item \code{phn_model()}: \link{PHN_MODEL}

}}
