% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_LOGNORMAL$..init..}
\alias{RAND_LOGNORMAL$..init..}
\title{Initialization method}
\arguments{
\item{mu}{Numeric. Mean of the log of the random variable.}

\item{sigma}{Numeric. Standard deviation of the log the random variable.}
}
\value{
Return the object itself.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{RAND_LOGNORMAL$..init..(mu = 0, sigma = 1)
}\if{html}{\out{</div>}}
}
}
\examples{

# Instantiate
test <- rand_lognormal(mu = 1, sigma = 2)
test
}
