% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\name{RAND_NORMAL$..init..}
\alias{RAND_NORMAL$..init..}
\title{Initialization method}
\arguments{
\item{mu}{Numeric.}

\item{sigma}{Numeric.}
}
\value{
No return value, called for side effects.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment.
}
\examples{

# Instantiation
test <- RAND_NORMAL$instantiation(mu = 1, sigma = 2)
test
}
