% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\docType{data}
\name{RAND_VAR}
\alias{RAND_VAR}
\title{RAND_VAR class environment}
\format{
An environment with S3 class \code{oop}.
}
\usage{
RAND_VAR
}
\description{
This is the base class of random variable, inherited from
\link{BASE}.
}
\seealso{
Parent class: \link{BASE}
\cr
\cr
New methods: \link{RAND_VAR$..init..}, \link{RAND_VAR$..str..},
\link{RAND_VAR$E}, \link{RAND_VAR$Var}
}
\keyword{datasets}
