% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_RAND_VAR.R
\docType{data}
\name{RAND_VAR}
\alias{RAND_VAR}
\title{RAND_VAR class environment}
\format{
An environment with S3 class \code{oop}.
}
\usage{
RAND_VAR
}
\description{
This is the base class of random variable.
}
\keyword{datasets}
