% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{sim_dist}
\alias{sim_dist}
\title{Approximate the distribution of number of detections of a lineup with simulation}
\usage{
sim_dist(n_evaluations, n_selections, n_plots = 20, n_simulations = 50000)
}
\arguments{
\item{n_evaluations}{Integer. Number of evaluations.}

\item{n_selections}{Integer. A vector of the number of selections.}

\item{n_plots}{Integer. Number of plots in the lineup.}

\item{n_simulations}{Integer. Number of simulations draws.}
}
\value{
A named vector representing the probability mass function of the distribution.
}
\description{
This function approximate the distribution of number of detections of a
lineup for given number of evaluations, selections in each evaluation and
plots in a lineup.
}
\details{
For a given lineup, plots are assumed to have weights
\eqn{W_i, i = 1, ..., N,} where \eqn{N} is the number of plots, and
\eqn{W_i} follows a uniform distribution independently. For each draw,
weights for a lineup will be simulated. Then, for each evaluation of a draw,
the function will sample same number of plots as the number of selection in
the evaluation using the simulated weights without replacement.
Finally, the distribution of the occurrences of plot 1 in a draw is the
approximated distribution of number of detections of a lineup.
}
\examples{
sim_dist(3, c(2,2,3))
sim_dist(1, c(1))

}
