% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$..cache..}
\alias{VI_MODEL$..cache..}
\title{Cache list, containing the last fitted model, data frame and formula}
\description{
A list, will be used if \code{cache = TRUE} while calling the \code{fit}
method.
}
\examples{

# Instantiate
x <- bandicoot::rand_uniform()
e <- bandicoot::rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

# Turn on cache
test$fit(test$gen(10), cache = TRUE)

test$..cache..
}
