% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$effect_size}
\alias{VI_MODEL$effect_size}
\title{Compute the effect size of the simulated data or the defined model}
\arguments{
\item{n}{Number of observations in a plot.}

\item{tol}{Tolerance of the convergence.}

\item{window_size}{Window size of the convergence check.}
}
\value{
The estimated effect size.
}
\description{
This function computes the effect size of the defined model.
}
