% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$formula}
\alias{VI_MODEL$formula}
\title{Closed form expression of \code{y}}
\description{
A formula, will be passed to \code{CLOSED_FORM$instantiation} to
define a closed form expression for \code{y}.
}
\examples{

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

# A formula with environment equals to `test`
test$formula

# `y` has the same expression
test$prm$y
}
