% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$null_formula}
\alias{VI_MODEL$null_formula}
\title{Formula for fitting the null model}
\description{
A quoted formula, will be used to fit the null model.
}
\examples{

# Instantiate
x <- bandicoot::rand_uniform()
e <- bandicoot::rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y ~ 1 + x + x^2 + e,
                 null_formula = y ~ x,
                 alt_formula = y ~ x + I(x^2))

# A formula with environment equals to `test`
test$null_formula

# By default, `fit` use the null formula
test$fit(test$gen(10))

# F-test also needs to use the null model
test$test(test$gen(1000))
}
