% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\name{VI_MODEL$prm}
\alias{VI_MODEL$prm}
\title{List of parameters}
\description{
A list, will be initialized after an instance is built.
}
\examples{

# Instantiation
x <- rand_uniform()
e <- rand_normal()
test <- vi_model(prm = list(x = x, e = e),
                 prm_type = list(x = "r", e = "r"),
                 formula = y~x+x^2+e)
test$prm
}
