% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_VI_MODEL.R
\docType{data}
\name{VI_MODEL}
\alias{VI_MODEL}
\title{VI_MODEL class environment}
\format{
An environment with S3 class \code{bandicoot_oop}.
}
\usage{
VI_MODEL
}
\description{
This is the base class of visual inference linear model,
inherited from \link[bandicoot:BASE]{bandicoot::BASE}.
}
\seealso{
Parent class: \link[bandicoot:BASE]{bandicoot::BASE}
\cr
\cr
New attributes: \link{VI_MODEL$prm}, \link{VI_MODEL$prm_type}, \link{VI_MODEL$..cache..},
\link{VI_MODEL$formula}, \link{VI_MODEL$null_formula}, \link{VI_MODEL$alt_formula}
\cr
\cr
New methods: \link{VI_MODEL$..init..}, \link{VI_MODEL$..str..}, \link{VI_MODEL$set_formula},
\link{VI_MODEL$gen}, \link{VI_MODEL$test}, \link{VI_MODEL$fit}, \link{VI_MODEL$effect_size},
\link{VI_MODEL$plot_resid},
\link{VI_MODEL$plot_qq},
\link{VI_MODEL$plot}, \link{VI_MODEL$plot_lineup}, \link{VI_MODEL$rss},
\link{VI_MODEL$null_resid}, \link{VI_MODEL$gen_lineup}
}
\keyword{datasets}
